Eliminated redundant methods in UnoHand
and UnoCard. Built and Tested Test classes for Unodeck, UnoCard. Built
but not yet tested Test class for UnoHand.
Fixed cascade of errors after above changes in UnoPlayer and UnoV2.
Got around 'cannot use non-static method in static context' for methods
UnoSpecialCardsV2 and for scoreUnoHand by instantiating a member of each
class where I needed it. Not sure this is a permanent fix...